+++ include/salmon.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#define ClearScreen() glClear(GL_COLOR_BUFFER_BIT); glClear(GL_DEPTH_BUFFER_BIT);
+
+#include <utils.h>
+#include <window.h>
+#include <input.h>
+#include <camera.h>
+#include <shader.h>
+#include <mesh.h>
+#include <model.h>
+#include <ecs.h>
+#include <stb_image.h>
+#include <engine.h>
+#include <components.h>
+#include <renderer.h>