+++ src/systems.cpp
@@ -1,0 +1,61 @@
+#include <components.h>
+#include <ecs.h>
+#include <engine.h>
+#include <iostream>
+#include <renderer.h>
+#include <GLFW/glfw3.h>
+
+void MeshRendererSys()
+{
+    for (EntityID ent : SceneView<Transform, MeshRenderer>(engineState.scene))
+    {
+        float aspectRatio = engineState.window->GetAspectRatio();
+        Renderer::RenderModel(ent, engineState.camera->GetProjMatrix(aspectRatio), engineState.camera->GetViewMatrix());
+    }
+}
+
+void CameraMoveSys()
+{
+    float deltaTime = 0.016f;
+    if (glfwGetKey(engineState.window->window, GLFW_KEY_W) == GLFW_PRESS)
+        engineState.camera->ProcessKeyboard(CameraMovement::FORWARD, deltaTime);
+    if (glfwGetKey(engineState.window->window, GLFW_KEY_S) == GLFW_PRESS)
+        engineState.camera->ProcessKeyboard(CameraMovement::BACKWARD, deltaTime);
+    if (glfwGetKey(engineState.window->window, GLFW_KEY_A) == GLFW_PRESS)
+        engineState.camera->ProcessKeyboard(CameraMovement::LEFT, deltaTime);
+    if (glfwGetKey(engineState.window->window, GLFW_KEY_D) == GLFW_PRESS)
+        engineState.camera->ProcessKeyboard(CameraMovement::RIGHT, deltaTime);
+}
+
+void CameraLookSys()
+{
+    // Variables to store the last mouse position
+    static bool firstMouse = true;
+    static float lastX = 0.0f, lastY = 0.0f;
+
+    // Get the current mouse position
+    double xpos, ypos;
+    glfwGetCursorPos(engineState.window->window, &xpos, &ypos);
+
+    if (firstMouse)
+    {
+        lastX = static_cast<float>(xpos);
+        lastY = static_cast<float>(ypos);
+        firstMouse = false;
+    }
+
+    // Calculate the offset from the last position
+    float xoffset = static_cast<float>(xpos) - lastX;
+    float yoffset = lastY - static_cast<float>(ypos); // Reversed since y-coordinates go from bottom to top
+
+    // Update the last position
+    lastX = static_cast<float>(xpos);
+    lastY = static_cast<float>(ypos);
+
+    // Process the mouse movement with the calculated offset
+    engineState.camera->ProcessMouseMovement(xoffset, yoffset);
+}
+
+REGISTER_SYSTEM(CameraLookSys);
+REGISTER_SYSTEM(MeshRendererSys);
+REGISTER_SYSTEM(CameraMoveSys);